requirements.txt
flask
flask_socketio
resemblyzer
numpy
sounddevice
whisper
torch
transformers
sklearn.cluster